Module3 Task5(edit2)

+++ src/main/java/lang/print/gaps/task5/TimeConvertor.java
@@ -1,7 +1,7 @@
 package lang.print.gaps.task5;
 
 public class TimeConvertor {
-    public void convert(float minutes) {
+    public void convert(int minutes) {
 
         minutes = minutes * 60;
         System.out.print(minutes);
